requirements.txt
fastapi
uvicorn
requests
beautifulsoup4
python-telegram-bot
sqlalchemy
apscheduler
playwright
python-dotenv
